backend/requirements.txt
flask
flask_cors
DateTime
# openai
gunicorn
python-dotenv
firebase-admin
Flask-PyMongo==2.3.0
PyMongo==4.6.1

pytest
pytest-mock
# redis
# Flask-Session